+++ RELEASE_NOTES.md
@@ -19,6 +19,7 @@
 - Clarify gender naming (4.0.0)
 - Fixed player update (4.0.1)
 - Fixed player search keyboard selection (4.0.1)
+- Invalidate year of birth < 1900 (4.0.2)
 
 ## Pairings